Split out util to be more testable (#108)
