+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export type * from './commonTypes';
 export type * from './serviceTypes';
 export type * from './treeViewTypes';
+export type * from './uiTypes';
