+++ src/types/uiTypes.d.ts
@@ -1,0 +1,18 @@
+import * as vscode from 'vscode';
+import type { ServerState } from './commonTypes';
+import type { IDhService } from './serviceTypes';
+
+export type SeparatorPickItem = {
+  label: string;
+  kind: vscode.QuickPickItemKind.Separator;
+};
+
+export type ConnectionPickItem<TType, TData> = vscode.QuickPickItem & {
+  type: TType;
+  data: TData;
+};
+
+export type ConnectionPickOption =
+  | SeparatorPickItem
+  | ConnectionPickItem<'server', ServerState>
+  | ConnectionPickItem<'connection', IDhService>;
